fixed bug in parts nav. removed old fonts. added latest font lib

+++ stage.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-rsync --delete -r dist/* johanrox@demo.spraakdata.gu.se:/export/htdocs_littb
+grunt && rsync --delete -r dist/* johanrox@demo.spraakdata.gu.se:/export/htdocs_littb
